requirements.txt
pytesseract
pillow
python-dotenv
transformers
torch
requests